Use distroless/static as base image
This is to make scanning tools happy, that only know about a limited
number of distros (b/366296183). Previously the image wasn't a proper
OCI because it lacked os-release and such, but that also isn't the root
of the problem. This new base image has ca-certificates, /tmp, and
tzdata, none of which do we actually want to use. This appears to be the
lowest-dependency base image understood by the scanning tool.